+++ packages/browser/tests/unit/browser.spec.ts
@@ -102,19 +102,6 @@ describe("go", () => {
   run_suite(go_suite);
 });
 
-describe("browser history.navigate", () => {
-  it("throws if trying to navigate with a non-encoded pathname", () => {
-    with_dom({ url: "http://example.com/one" }, ({ window }) => {
-      const test_history = browser(pending => {
-        pending.finish();
-      });
-      expect(() => {
-        test_history.navigate("/test ing");
-      }).toThrow();
-    });
-  });
-});
-
 describe("browser history.go", () => {
   // integration?
   it("calls window.history.go with provided value", () => {