+++ packages/dom-utils/types/index.d.ts
@@ -1,4 +1,3 @@
-export declare function verify_encoded_pathname(pathname: string): boolean;
 export declare function dom_exists(): boolean;
 export declare function ignorable_popstate_event(event: PopStateEvent): boolean;
 export declare function get_state_from_history(): any;