+++ packages/hash/src/hash.ts
@@ -1,9 +1,5 @@
 import { location_utils, key_generator, navigate_with } from "@hickory/root";
-import {
-  get_state_from_history,
-  dom_exists,
-  verify_encoded_pathname
-} from "@hickory/dom-utils";
+import { get_state_from_history, dom_exists } from "@hickory/dom-utils";
 import hash_encoder_and_decoder from "./hashTypes";
 
 import {
@@ -32,10 +28,6 @@ export function hash(
 ): HashHistory {
   if (!dom_exists()) {
     throw new Error("Cannot use @hickory/hash without a DOM");
-  }
-
-  if (!options.pathname) {
-    options.pathname = verify_encoded_pathname;
   }
 
   const location_utilities = location_utils(options);